Adicionando regas da contratação plano e resgate

+++ prev/routes/resgate.py
@@ -1,10 +1,11 @@
-import uuid
+from datetime import datetime
 from typing import List
 from fastapi import APIRouter, status
 from fastapi.exceptions import HTTPException
 from sqlmodel import Session, select
 
 from prev.routes import crud
+from prev.utils import calcula_data_carencia
 from prev.db import ActiveSession
 from prev.models.user import Resgate
 from prev.serializers import ResgateRequest, ResgateIdResponse
@@ -26,18 +27,22 @@ async def create_resgate(*, session: Session = ActiveSession, resgate_request: R
     """Criar novos resgates"""
 
     db_plano = crud.get_plano(db=session, plano_id=resgate_request.idPlano)
+    db_produto = crud.get_produto(db=session, producto_id=db_plano.idProduto)
 
     if not db_plano:
         raise HTTPException(status_code=404, detail="Plano não encontrado")
-    
-    #TODO: adicionar validação do prazo de resgate
-    
+
+    data_atual = datetime.now()
+    dias_diferenca = calcula_data_carencia(db_plano.dataDaContratacao, data_atual)
+
+    if dias_diferenca < db_produto.carenciaInicialDeResgate:
+        return dias_diferenca
+
     if db_plano.aporte < resgate_request.valorResgate:
         raise HTTPException(status_code=400, detail="Saldo insuficiente para resgate.")
-    
-    #db_produto = crud.get_produto(db=session, producto_id=db_plano.idProduto)
 
     db_plano.aporte -= resgate_request.valorResgate
+
     if db_plano.aporte == 0:
         session.delete(db_plano)
